assignment-2/requirements.txt
kafka-python
faker
mysql-connector-python
